First item and new random PhysCircle class
Implemented a basic item pickup, and added a class to hold a "Loot Pool" which returns one of many PhysCircles given to it, based on a "chance" value given to each one in the list. This is a universal random PhysCircle retriever, so it can be used with any GameObject to create random items, bombs, projectiles, etc.